plant_id_project/requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
requests==2.31.0
openai==1.30.1
line-bot-sdk==3.3.0
gunicorn==21.2.0
Werkzeug==2.3.7
pip install Flask==2.3.2 python-dotenv==1.0.1 mysql-connector-python==8.3.0 \
    requests==2.31.0 openai==1.30.1 line-bot-sdk==3.3.0 gunicorn==21.2.0 Werkzeug==2.3.7

.env.example
PLANTNET_API_KEY=your-plantnet-key
OPENAI_API_KEY=your-openai-key
SECRET_KEY=your-flask-secret
DB_HOST=localhost
DB_PORT=3306
DB_USER=root
DB_PASSWORD=yourpassword
DB_NAME=yourdbname
